Fix deletion participant test wiring
