+++ anyharness/crates/anyharness-lib/src/sessions/links/service.rs
@@ -323,280 +323,5 @@ fn is_unique_constraint_error(error: &anyhow::Error) -> bool {
 }
 
 #[cfg(test)]
-mod tests {
-    use rusqlite::params;
-
-    use super::*;
-    use crate::{
-        persistence::Db, sessions::deletion::SessionDeleteWorkflow, sessions::model::SessionRecord,
-    };
-
-    fn seed_workspace(db: &Db) {
-        db.with_conn(|conn| {
-            conn.execute(
-                "INSERT INTO workspaces (id, kind, path, source_repo_root_path, created_at, updated_at)
-                 VALUES (?1, 'repo', '/tmp/workspace', '/tmp/workspace', ?2, ?2)",
-                params!["workspace-1", "2026-03-25T00:00:00Z"],
-            )?;
-            Ok(())
-        })
-        .expect("seed workspace");
-    }
-
-    fn session_record(id: &str) -> SessionRecord {
-        SessionRecord {
-            id: id.to_string(),
-            workspace_id: "workspace-1".to_string(),
-            agent_kind: "claude".to_string(),
-            native_session_id: None,
-            agent_auth_scope: None,
-            required_agent_auth_revision: None,
-            requested_model_id: None,
-            current_model_id: None,
-            requested_mode_id: None,
-            current_mode_id: None,
-            title: None,
-            thinking_level_id: None,
-            thinking_budget_tokens: None,
-            status: "idle".to_string(),
-            created_at: "2026-03-25T00:00:00Z".to_string(),
-            updated_at: "2026-03-25T00:00:00Z".to_string(),
-            last_prompt_at: None,
-            closed_at: None,
-            dismissed_at: None,
-            mcp_bindings_ciphertext: None,
-            mcp_binding_summaries_json: None,
-            mcp_binding_policy: crate::sessions::model::SessionMcpBindingPolicy::InheritWorkspace,
-            system_prompt_append: None,
-            subagents_enabled: true,
-            action_capabilities_json: None,
-            origin: None,
-        }
-    }
-
-    fn service_fixture() -> (Db, SessionStore, SessionLinkService) {
-        let db = Db::open_in_memory().expect("open db");
-        seed_workspace(&db);
-        let session_store = SessionStore::new(db.clone());
-        session_store
-            .insert(&session_record("parent-1"))
-            .expect("insert parent");
-        session_store
-            .insert(&session_record("child-1"))
-            .expect("insert child");
-        let service =
-            SessionLinkService::new(SessionLinkStore::new(db.clone()), session_store.clone());
-        (db, session_store, service)
-    }
-
-    fn create_input(parent: &str, child: &str) -> CreateSessionLinkInput {
-        CreateSessionLinkInput {
-            relation: SessionLinkRelation::Subagent,
-            parent_session_id: parent.to_string(),
-            child_session_id: child.to_string(),
-            workspace_relation: SessionLinkWorkspaceRelation::SameWorkspace,
-            label: Some("Child".to_string()),
-            created_by_turn_id: Some("turn-1".to_string()),
-            created_by_tool_call_id: Some("tool-1".to_string()),
-        }
-    }
-
-    #[test]
-    fn creates_and_lists_links_by_parent_and_child() {
-        let (_db, _session_store, service) = service_fixture();
-
-        let link = service
-            .create_link(create_input("parent-1", "child-1"))
-            .expect("create link");
-
-        let by_parent = service.list_by_parent("parent-1").expect("list by parent");
-        assert_eq!(by_parent, vec![link.clone()]);
-        let by_child = service.list_by_child("child-1").expect("list by child");
-        assert_eq!(by_child, vec![link]);
-    }
-
-    #[test]
-    fn rejects_missing_parent_or_child() {
-        let (_db, _session_store, service) = service_fixture();
-
-        let missing_parent = service
-            .create_link(create_input("missing-parent", "child-1"))
-            .expect_err("missing parent");
-        assert!(matches!(
-            missing_parent,
-            CreateSessionLinkError::ParentNotFound(id) if id == "missing-parent"
-        ));
-
-        let missing_child = service
-            .create_link(create_input("parent-1", "missing-child"))
-            .expect_err("missing child");
-        assert!(matches!(
-            missing_child,
-            CreateSessionLinkError::ChildNotFound(id) if id == "missing-child"
-        ));
-    }
-
-    #[test]
-    fn rejects_self_links_and_duplicates() {
-        let (_db, _session_store, service) = service_fixture();
-
-        let self_link = service
-            .create_link(create_input("parent-1", "parent-1"))
-            .expect_err("self link");
-        assert!(matches!(self_link, CreateSessionLinkError::SelfLink));
-
-        service
-            .create_link(create_input("parent-1", "child-1"))
-            .expect("create first link");
-        let duplicate = service
-            .create_link(create_input("parent-1", "child-1"))
-            .expect_err("duplicate link");
-        assert!(matches!(duplicate, CreateSessionLinkError::Duplicate));
-    }
-
-    #[test]
-    fn rejects_second_subagent_parent_for_same_child() {
-        let (_db, session_store, service) = service_fixture();
-        session_store
-            .insert(&session_record("parent-2"))
-            .expect("insert second parent");
-
-        service
-            .create_link(create_input("parent-1", "child-1"))
-            .expect("create first link");
-        let duplicate_parent = service
-            .create_link(create_input("parent-2", "child-1"))
-            .expect_err("second parent rejected");
-
-        assert!(matches!(
-            duplicate_parent,
-            CreateSessionLinkError::ChildAlreadyLinked
-        ));
-    }
-
-    #[test]
-    fn create_subagent_link_enforces_child_limit_at_insert() {
-        let (_db, session_store, service) = service_fixture();
-        session_store
-            .insert(&session_record("child-2"))
-            .expect("insert second child");
-
-        service
-            .create_subagent_link_with_child_limit(create_input("parent-1", "child-1"), 1)
-            .expect("create first child link");
-        let limit = service
-            .create_subagent_link_with_child_limit(create_input("parent-1", "child-2"), 1)
-            .expect_err("fanout limit");
-
-        assert!(matches!(limit, CreateSessionLinkError::FanoutLimit));
-        assert!(service
-            .find_subagent_link("parent-1", "child-2")
-            .expect("find second child link")
-            .is_none());
-    }
-
-    #[test]
-    fn closed_links_are_hidden_from_normal_lists_but_available_to_history() {
-        let (_db, _session_store, service) = service_fixture();
-        let link = service
-            .create_link(create_input("parent-1", "child-1"))
-            .expect("create link");
-
-        service
-            .close_link(&link.id, "2026-03-25T00:02:00Z")
-            .expect("mark closed");
-
-        assert!(service
-            .list_by_parent("parent-1")
-            .expect("list open by parent")
-            .is_empty());
-        assert!(service
-            .list_by_child("child-1")
-            .expect("list open by child")
-            .is_empty());
-        let historical = service
-            .list_by_parent_including_closed("parent-1")
-            .expect("list historical by parent");
-        assert_eq!(historical.len(), 1);
-        assert_eq!(historical[0].id, link.id);
-        assert_eq!(
-            historical[0].closed_at.as_deref(),
-            Some("2026-03-25T00:02:00Z"),
-        );
-    }
-
-    #[test]
-    fn import_link_backfills_missing_public_id() {
-        let (_db, session_store, service) = service_fixture();
-        session_store
-            .insert(&session_record("child-2"))
-            .expect("insert imported child");
-        let mut record = service
-            .create_link(create_input("parent-1", "child-1"))
-            .expect("create template link");
-        record.id = "imported-link".to_string();
-        record.public_id = None;
-        record.child_session_id = "child-2".to_string();
-
-        service.import_link(&record).expect("import link");
-
-        let imported = service
-            .list_by_child("child-2")
-            .expect("list imported child")
-            .pop()
-            .expect("imported link");
-        assert_eq!(imported.id, "imported-link");
-        assert!(imported
-            .public_id
-            .as_deref()
-            .is_some_and(|id| id.starts_with("subagent_")));
-    }
-
-    #[test]
-    fn rejects_unknown_enum_values_on_read() {
-        let (db, _session_store, service) = service_fixture();
-        db.with_conn(|conn| {
-            conn.execute(
-                "INSERT INTO session_links (
-                    id, relation, parent_session_id, child_session_id, workspace_relation,
-                    created_at
-                 ) VALUES ('bad-relation', 'sidekick', 'parent-1', 'child-1', 'same_workspace', ?1)",
-                ["2026-03-25T00:01:00Z"],
-            )?;
-            Ok(())
-        })
-        .expect("insert bad relation");
-        assert!(service.list_by_parent("parent-1").is_err());
-
-        db.with_conn(|conn| {
-            conn.execute("DELETE FROM session_links", [])?;
-            conn.execute(
-                "INSERT INTO session_links (
-                    id, relation, parent_session_id, child_session_id, workspace_relation,
-                    created_at
-                 ) VALUES ('bad-workspace-relation', 'subagent', 'parent-1', 'child-1', 'new_worktree', ?1)",
-                ["2026-03-25T00:01:00Z"],
-            )?;
-            Ok(())
-        })
-        .expect("insert bad workspace relation");
-        assert!(service.list_by_parent("parent-1").is_err());
-    }
-
-    #[test]
-    fn delete_session_removes_parent_and_child_links() {
-        let (db, _session_store, service) = service_fixture();
-        service
-            .create_link(create_input("parent-1", "child-1"))
-            .expect("create link");
-
-        SessionDeleteWorkflow::new(db)
-            .delete_session("parent-1")
-            .expect("delete parent");
-
-        assert!(service
-            .list_by_child("child-1")
-            .expect("list by child")
-            .is_empty());
-    }
-}
+#[path = "service_tests.rs"]
+mod tests;
